Add example to README.md
For this commit a new file example.txt was added with the text:

    This is the first line.

    This is the fifth line.

The example shows the contents of this file (with cat) and then it feeds
to dss the file, redirecting the output to example.out. Then example.out
is shown which obviously has incorrect text (fifth will be on the third
line) so we use sed again (with -i'' [0]) to fix the text in example.out
Finally we show the corrected file. Of course one can use a more
advanced command (or commands) but this shows the original state and how
the script works.

.gitignore added with example.out as this should not be in git.

[0] We use -i'' and not -i because not all implementations of sed allow
the -i option without an extension but an empty string is allowed in
these implementation (or at least all I have seen). The empty string is,
as far as I am aware, a GNU extension; in BSD it is not supported for
instance (macOS does not support it and iirc from ages past
SunOS/Solaris versions do not support it either - I don't recall FreeBSD
but I doubt it supports it unless you install the GNU sed and have it in
front of the rest of your path).